services/etl/requirements.txt
python-dotenv==1.1.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.11.4
pydantic_settings==2.9.1
requests==2.32.3
redis==5.2.1
